+++ src/containers/AddMembersToList/Entities/index.js
@@ -1,15 +1,12 @@
 import { connect } from 'react-redux';
 import AddMemberToListLayout from '../layout';
-import {
-  setSelectedSubEntityId,
-  toggleListItemEntity
-} from '../../../redux/modules/entities';
+import { setSelectedSubEntityId, toggleListItemEntity } from '../../../redux/modules/entities';
 import {
   userHasUpdatePermission,
   getCurrentEntity,
   getSelectedSubEntityId
 } from '../../../redux/modules/entities/selectors';
-import { getModalTableItems } from '../../../redux/modules/entities/listItemSelectors';
+import { selectModalTableItems } from './selectors';
 import { entitiesMetaData } from '../../../redux/modules/entities/metaData';
 
 export function mapStateToProps(state) {
@@ -18,13 +15,9 @@ export function mapStateToProps(state) {
     userHasUpdatePermission: userHasUpdatePermission(state),
     listName: getSelectedSubEntityId(state),
     entityName: currentEntity,
-    fields:
-      entitiesMetaData[getSelectedSubEntityId(state)].memberListTableFields,
-    tableItems: getModalTableItems(state, getSelectedSubEntityId(state)),
-    defaultFilters:
-      entitiesMetaData[currentEntity].defaultAssociationFilters[
-        getSelectedSubEntityId(state)
-      ],
+    fields: entitiesMetaData[getSelectedSubEntityId(state)].memberListTableFields,
+    tableItems: selectModalTableItems(state, getSelectedSubEntityId(state)),
+    defaultFilters: entitiesMetaData[currentEntity].defaultAssociationFilters[getSelectedSubEntityId(state)],
     contains: getSelectedSubEntityId(state)
   };
 }