+++ src/redux/modules/entities/skills/selectors.js
@@ -5,16 +5,19 @@
 import { getSelectedEntityId } from '../selectors';
 import { getDependantEntityTableItems } from '../listItemSelectors';
 import { getCurrentForm } from '../../form/selectors';
+import { getDisplay } from '../users/selectors';
 
 export const getHasProficiencyFormValue = state =>
-  getCurrentForm(state) && getCurrentForm(state).getIn(['values', 'hasProficiency']);
+  getCurrentForm(state) &&
+  getCurrentForm(state).getIn(['values', 'hasProficiency']);
 
 export const getSkillMemberSidePanelTableItems = state => {
   const selectedEntityId = getSelectedEntityId(state);
   return getDependantEntityTableItems(state).map(user => ({
     ...user,
-    firstName: user.firstName !== null ? user.firstName : '',
-    lastName: user.lastName !== null ? user.lastName : '',
-    proficiency: user.skills.filter(skill => selectedEntityId === skill.id).map(skill => skill.proficiency)[0]
+    name: getDisplay(user),
+    proficiency: user.skills
+      .filter(skill => selectedEntityId === skill.id)
+      .map(skill => skill.proficiency)[0]
   }));
 };